requirements.txt
aiogram==3.4.1
openai==0.28.1
PyGithub==2.1.1
GitPython==3.1.42
pydantic>=2.4.1,<2.6
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.5
unidiff==0.7.5
structlog==24.1.0
cryptography==41.0.7
PyJWT==2.8.0
httpx==0.24.1
